Handle correctly signal from delayed X11 start
Handling of the SIGUSR1 signal after X11 start timed out was wrong. Normally,
this signal is used to test the exception handler. If X11 timed out but
eventually started and sent SIGUSR1 to Anaconda, the signal was interpreted as
testing the exception handler and Anaconda crashed.

With this change, the X1-specific handler for SIGUSR1 is kept until the first
SIGUSR1 arrives.

This changes the possible corner case. Now, if X11 truly fails, invoking the
exception testing needs SIGUSR1 twice.

Resolves: rhbz#1918702